Add code formatting check to Travis CI build
Check compliance with the code style standard established by Arduino's examples_formatter.conf Artistic Style configuration file:
https://github.com/arduino/Arduino/blob/master/build/shared/examples_formatter.conf
If the code does not comply with the standard, a diff between the submitted code and the auto-formatted code is shown in the log and the Travis CI job will fail.

I configured .travis.yml so that the formatting check will be done in a dedicated job.