Check for commonly misspelled words during the CI build
codespell is a blacklist-based spell checker. This approach is more appropriate for a CI build because it results in far less false positives than would traditional whitelist-based spell checking, with the disadvantage of having more false negatives. So it won't catch all the typos, but it also won't be spuriously breaking every other build.

When false positives do occur, they can be resolved by adding the problematic word to extras/codespell-ignore-words-list.txt

I configured codespell to ignore the contents of src/lib and test/external because these appear to contain external projects and spell checking of those files would be better done in their own repositories.